+++ CHANGELOG.md
@@ -4,14 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.4.0] - 2020-10-13
 ### Added
 - `JewishCalendar#end_of_week` returns Shabbos for this week
 - Parshas Hashavua modifier to return blank Limud if the standard Parsha is not read this Shabbos
 - Daf Hashavua calculator
+### Changed
 - Refactor kviah calculation logic to JewishDate
 - Remove formatter dependency for Parsha calculator
-
 ### Fixed
 - Added Lag B'Omer as significant day / yom tov
 - Fixed Lag B'Omer spelling for english formatted omer